qubesu2f/tools/u2fdump: fix SIGINT handling
CancelledError thrown into loop.run_in_executor() gets raised back only
on next iteration, so when ioctl was blocked there, it didn't really
cancel the coroutine. Now this is done with loop's reader, with a catch
that selectors don't generally work with ioctl(), it is only by chance
that this particular character device has also read() defined which
works almost exactly like the ioctl in question.